lambda/cellsegmentation_image/requirements.txt
boto3
Pillow
numpy
matplotlib
scikit-image